+++ ad-tag/index.ts
@@ -4,6 +4,7 @@ export * from './source/ts/types/apstag';
 export * from './source/ts/types/googletag';
 export * from './source/ts/types/moli';
 export * from './source/ts/types/module';
+export * from './source/ts/types/tcfapi';
 export * from './source/ts/ads/moliGlobal';
 export * from './source/ts/ads/moli';
 export * from './source/ts/ads/adPipeline';