+++ examples/ie-11/source/ts/configuration.ts
@@ -192,7 +192,6 @@ export const adConfiguration: Moli.MoliConfig = {
       }
     }
   },
-  consent: {},
   yieldOptimization: {
     provider: 'none',
   },