+++ examples/publisher/source/ts/configuration.ts
@@ -470,7 +470,6 @@ export const adConfiguration: Moli.MoliConfig = {
     cmpTimeout: 500,
     pubID: 'test'
   },
-  consent: {},
   yieldOptimization: {
     provider: 'dynamic',
     configEndpoint: '//local.h5v.eu:9000/yield-config.json'