+++ examples/single-page-app/index.ts
@@ -10,19 +10,13 @@ import 'prebid.js/modules/appnexusBidAdapter';
 
 import { initAdTag } from '@highfivve/ad-tag';
 import { adConfiguration } from './source/ts/configuration';
-import Faktor from '@highfivve/module-cmp-faktor';
+import SourcepointCmp from '@highfivve/module-cmp-sourcepoint';
 
 prebid.processQueue();
 
 const moli = initAdTag(window);
 
-moli.registerModule(new Faktor({
-  autoOptIn: true,
-  site: {
-    mode: 'lazy',
-    url: 'https://config-prod.choice.faktor.io/cb5df6d3-99b4-4d5b-8237-2ff9fa97d1a0/faktor.js'
-  }
-}, window));
+moli.registerModule(new SourcepointCmp(window));
 
 moli.enableSinglePageApp();
 // init moli