+++ modules/faktor-cmp/faktorStub.d.ts
@@ -1,2 +1,0 @@
-declare function loadCmpFaktorStub(window: Window): void;
-export = loadCmpFaktorStub;