+++ modules/generic-cmp/cmpStub.d.ts
@@ -1,2 +1,0 @@
-declare function loadStub(window: Window): void;
-export = loadStub;