+++ moli-debugger/components/consentConfig.tsx
@@ -2,21 +2,30 @@ import * as preact from 'preact';
 
 import { Tag } from './tag';
 
-import { Moli } from '@highfivve/ad-tag';
-import { ConsentString } from 'consent-string';
-import { IABConsentManagement } from '@highfivve/ad-tag/source/ts/types/IABConsentManagement';
-import IConsentData = IABConsentManagement.IConsentData;
+import { tcfapi } from '@highfivve/ad-tag';
+import { TCModel, TCString } from '@iabtcf/core';
 
-type IConsentConfigProps = {
-  consent: Moli.consent.ConsentConfig
+type CmpVendor = {
+  name: string;
+  website: string;
+  subDomain: string;
+};
+type CmpVendors = { [id: string]: CmpVendor | undefined };
+
+const cmpVendors: CmpVendors = {
+  '3': { name: ' LiveRamp', website: 'https://liveramp.com/', subDomain: '\tfaktor.mgr.consensu.org' },
+  '5': { name: ' Usercentrics GmbH', website: 'https://usercentrics.com/', subDomain: 'usercentrics.mgr.consensu.org' },
+  '6': { name: 'Sourcepoint Technologies, Inc.', website: 'https://sourcepoint.com', subDomain: 'sourcepoint.mgr.consensu.org' },
+  '10': { name: 'Quantcast International Limited', website: 'https://quantcast.com', subDomain: 'quantcast.mgr.consensu.org' },
+  '21': { name: 'Traffective GmbH', website: 'https://traffective.com', subDomain: 'traffective.mgr.consensu.org' },
 };
 
+declare const window: Window & tcfapi.TCFApiWindow;
+
 type IConsentConfigState = {
-  consentData?: IConsentData,
-  nonPersonalizedAds?: 0 | 1,
-  vendorListVersion?: number,
-  numAllowedVendors?: number,
-  allowedPurposes: number[],
+  cmpStatus?: tcfapi.status.CmpStatus,
+  tcModel?: TCModel,
+  tcString?: string,
   messages: Message[];
 };
 
@@ -25,62 +34,76 @@ type Message = {
   text: string | JSX.Element;
 };
 
-export class ConsentConfig extends preact.Component<IConsentConfigProps, IConsentConfigState> {
+export class ConsentConfig extends preact.Component<{}, IConsentConfigState> {
 
-  constructor(props: IConsentConfigProps) {
+  constructor() {
     super();
     this.state = {
-      allowedPurposes: [],
       messages: []
     };
-    this.checkConsentConfig(this.state.messages, props.consent);
-    this.initConsentData(this.state.messages, props.consent);
+    this.checkConsentConfig(this.state.messages);
+    this.initConsentData();
   }
 
-  render(props: IConsentConfigProps, state: IConsentConfigState): JSX.Element {
+  render(props: {}, state: IConsentConfigState): JSX.Element {
     return <div>
-      {this.consent(props.consent)}
+      {this.consent()}
       {this.consentData()}
     </div>;
   }
 
-  private consent = (consent: Moli.consent.ConsentConfig): JSX.Element => {
+  private consent = (): JSX.Element => {
+
     return <div>
       <div className="MoliDebug-tagContainer">
-        <span className="MoliDebug-tagLabel">Module</span>
-        {consent.cmp ? <Tag>{consent.cmp.name}</Tag> : <Tag variant="red">No defined</Tag>}
+        <span className="MoliDebug-tagLabel"><a href="https://iabeurope.eu/cmp-list/">CMP ID</a></span>
+        {this.cmpVendorTag()}
       </div>
       <div className="MoliDebug-tagContainer">
-        <span className="MoliDebug-tagLabel">Config</span>
-        {consent.cmp ? <Tag>{JSON.stringify(consent.cmp.config())}</Tag> : <Tag variant="red">No config</Tag>}
+        <span className="MoliDebug-tagLabel">Last updated</span>
+        {this.state.tcModel ? <Tag>{this.state.tcModel.lastUpdated.toLocaleString()}</Tag> :
+          <Tag variant="red">Unknown</Tag>}
+      </div>
+      <div className="MoliDebug-tagContainer">
+        <span className="MoliDebug-tagLabel">TCF Version</span>
+        {this.state.tcModel ? <Tag>{this.state.tcModel.version.toString()}</Tag> : <Tag variant="red">Unknown</Tag>}
       </div>
     </div>;
   };
 
-  private initConsentData = (messages: Message[], consent?: Moli.consent.ConsentConfig): void => {
-    if (consent && consent.cmp) {
-      consent.cmp.getConsentData().then(consentData => {
-        this.setState({ consentData });
-        const consentString = new ConsentString(consentData ? consentData.consentData : undefined);
-        this.setState({ vendorListVersion: consentString.getVendorListVersion() });
-        this.setState({ numAllowedVendors: consentString.getVendorsAllowed().length });
-        this.setState({ allowedPurposes: consentString.getPurposesAllowed() });
-      }).catch(error => {
-        messages.push({
-          kind: 'error',
-          text: `Could'nt get ConsentData. ${error}`
-        });
-      });
-
-      consent.cmp.getNonPersonalizedAdSetting().then(nonPersonalizedAds => {
-        this.setState({ nonPersonalizedAds });
-      }).catch(error => {
-        messages.push({
-          kind: 'error',
-          text: `Could'nt get nonPersonalizedAds setting. ${error}`
-        });
-      });
+  private cmpVendorTag = (): JSX.Element => {
+    if (this.state.tcModel) {
+      const cmpVendor = cmpVendors[this.state.tcModel.cmpId.toString()];
+      if (cmpVendor) {
+        return <Tag variant="grey"><a href={cmpVendor.website}>{cmpVendor.name} (id: {this.state.tcModel.cmpId.toString()})</a></Tag>;
+      } else {
+        return <Tag variant="yellow">{this.state.tcModel.cmpId.toString()}</Tag>;
+      }
+    } else {
+      return <Tag variant="red">No defined</Tag>;
     }
+  }
+
+  private initConsentData = (): void => {
+    // fetch initial TCData
+    window.__tcfapi('getTCData', 2, (data: tcfapi.responses.TCData) => {
+      const tcModel = TCString.decode(data.tcString);
+      this.setState({
+        cmpStatus: data.cmpStatus,
+        tcModel: tcModel,
+        tcString: data.tcString
+      });
+    });
+
+    // Update on changes
+    window.__tcfapi('addEventListener', 2, event => {
+      const tcModel = TCString.decode(event.tcString);
+      this.setState({
+        cmpStatus: event.cmpStatus,
+        tcModel: tcModel,
+        tcString: event.tcString
+      });
+    });
   };
 
   private consentData = (): JSX.Element | undefined => {
@@ -89,15 +112,30 @@ export class ConsentConfig extends preact.Component<IConsentConfigProps, IConsen
       return <div>
         <div class="MoliDebug-tagContainer">
           <span class="MoliDebug-tagLabel">vendor list version</span>
-          <Tag>{this.state.vendorListVersion ? this.state.vendorListVersion.toString() : 'not found'}</Tag>
+          <Tag>{this.state.tcModel ? this.state.tcModel.vendorListVersion.toString() : 'not found'}</Tag>
         </div>
         <div class="MoliDebug-tagContainer">
           <span class="MoliDebug-tagLabel">num vendors allowed</span>
-          <Tag>{this.state.numAllowedVendors ? this.state.numAllowedVendors.toString() : 'not found'}</Tag>
+          <Tag>{this.state.tcModel ? this.state.tcModel.vendorConsents.size.toString() : 'not found'}</Tag>
         </div>
         <div class="MoliDebug-tagContainer">
-          <span class="MoliDebug-tagLabel">allowed purposes</span>
-          <Tag>{this.state.allowedPurposes ? this.state.allowedPurposes.toString() : 'not found'}</Tag>
+          <span className="MoliDebug-tagLabelBtn" type="button" data-toggle="collapse"
+                data-target="#collapsePurposes"
+                aria-expanded="false" aria-controls="collapsePurposes">
+            allowed purposes
+          </span>
+          <Tag>{this.state.tcModel ? this.state.tcModel.purposeConsents.size.toString() : 'not found'}</Tag>
+
+          <div className="collapse" id="collapsePurposes">
+            <a
+              href="https://iabeurope.eu/iab-europe-transparency-consent-framework-policies/#Appendix_A_Purposes_and_Features_Definitions">Purpose
+              and Feature definitions</a>
+            <br/>
+            {this.state.tcModel && Array.from(this.state.tcModel.purposeConsents).map(([ id, accepted ]) => {
+              return <Tag
+                variant={accepted ? 'green' : 'red'}>{id.toString()}: {accepted ? 'accepted' : 'denied'}</Tag>;
+            })}
+          </div>
         </div>
 
         <div class="MoliDebug-tagContainer">
@@ -105,41 +143,22 @@ export class ConsentConfig extends preact.Component<IConsentConfigProps, IConsen
                 aria-expanded="false" aria-controls="collapseConsentString">
             consent string
           </span>
-          <Tag>
-            <a class="link"
-               href={`https://useless.af/consent-decoder?gdpr_consent=${this.state.consentData ? this.state.consentData.consentData : ''}`}>
-              useless.af
-            </a>
-          </Tag>
           <div class="collapse" id="collapseConsentString">
-            <Tag>{this.state.consentData ? this.state.consentData.consentData : 'not found'}</Tag>
+            <Tag>{this.state.tcString ? this.state.tcString : 'not found'}</Tag>
           </div>
         </div>
       </div>;
     }
   };
 
-  private isCmpFunctionAvailable = () => window.__cmp || typeof window.__cmp === 'function';
+  private isCmpFunctionAvailable = () => window.__tcfapi || typeof window.__tcfapi === 'function';
 
-  private checkConsentConfig = (messages: Message[], consent?: Moli.consent.ConsentConfig): void => {
-    if (!consent) {
-      messages.push({
-        kind: 'error',
-        text: 'No consent configuration found.'
-      });
-    }
-    if (consent && !consent.cmp) {
-      messages.push({
-        kind: 'error',
-        text: 'no cmp module configured!'
-      });
-    }
-
+  private checkConsentConfig = (messages: Message[]): void => {
     // if cmp is configured, there must be a cmp present
-    if (consent && consent.cmp && !this.isCmpFunctionAvailable()) {
+    if (!this.isCmpFunctionAvailable()) {
       messages.push({
         kind: 'error',
-        text: 'no window.__cmp function found. Consent management and ads will not work!'
+        text: 'no window.__tcfapi function found. Consent management and ads will not work!'
       });
     }
   };