+++ moli-debugger/components/tag.tsx
@@ -2,7 +2,7 @@ import * as preact from 'preact';
 
 import { classList } from '../util/stringUtils';
 
-type TagVariant = 'green' | 'red' | 'yellow' | 'blue';
+type TagVariant = 'green' | 'red' | 'yellow' | 'blue' | 'grey';
 type TagSpacing = 'medium';
 
 export const Tag = (props: { children: JSX.Element | string | string[], variant?: TagVariant, title?: string, spacing?: TagSpacing }): JSX.Element =>  {